Implemented the "original background" extra feature as a part of the split output feature.
It allows to preserve the original image background when BW foreground is used in the format ready for the further processing.
It can be used to encode into DjVu the pages with the complex background using the semi-auto "split layers" method which gives much higher quality results than DjVu auto segmentator.
Also this feature can be used to extract high contrast elements of gradient images into the foreground layer by using second processing of the layer with pictures ("background").

Properties of the original background:
* Original background images are saved into "original_background" folder in "out" directory.
* Pure black (#000000) and white (#ffffff) colors of original background image are reserved into #010101 and #fefefe, respectively.
* Picture zones are marked with black when the BW content is marked with white. This property allow to use "select by color" feature of an image editor to select needed areas for their further processing, for example, apply blur to white holes and their nearest areas to get an effective compression level of the background layer in DjVu.
* Filling zones feature also removes trash and speckles from the original background when applied to the foreground layer.