Implemented the feature of auto adjusting content area.
Use double-click on content to automatically adjust the content area.
If the content is outside the area, the later will automatically be expanded and adjusted to the content at the position where double-click has been,
otherwise the area edge, nearest to that position, will be adjusted.

It's much faster now to correct the content area if, for example, the page number has been missed by the auto algorithm.
It is no more required to manually and laboriously move the corners and edges of the content box.